🎨 rename vue files
