+++ src/router/routes/modules/web-home.ts
@@ -6,12 +6,12 @@ export const WebHomeRoute: AppRouteRecordRaw = {
     meta: {
         title: '主页',
     },
-    component: () => import('@/views/web/web-home.vue'),
+    component: () => import('@/views/web/WebHome.vue'),
     children: [
         {
             path: 'setting',
             name: 'web-setting',
-            component: () => import('@/views/Setting.vue'),
+            component: () => import('@/views/setting/Setting.vue'),
             meta: {
                 title: '设置',
             }
